move code out of context module
